menu_letter: single-precision cursorY conversion in LetterListDraw (82.35->82.51%)
The DrawCursor cursorY int-to-int-via-float conversion uses single
precision (fsubs) in the target, not double (fsub). Use static_cast<float>
to match. Values (s16*0x20+0x60) are within float exact-int range, so
behavior-identical.

+++ src/menu_letter.cpp
@@ -1842,7 +1842,7 @@ void CMenuPcs::LetterListDraw()
 	}
 
 	const int cursorX = static_cast<int>(FLOAT_803330f4 + static_cast<float>(static_cast<int>(System.m_frameCounter) % 8));
-	const int cursorY = static_cast<int>(static_cast<double>(*reinterpret_cast<s16*>(GetLetterStateBase(this) + 0x26) * 0x20 + 0x60));
+	const int cursorY = static_cast<int>(static_cast<float>(*reinterpret_cast<s16*>(GetLetterStateBase(this) + 0x26) * 0x20 + 0x60));
 	DrawCursor(cursorX, cursorY, FLOAT_803330f8);
 }
